requirements.txt
aiofiles==0.8.0
py-avataaars==1.1.2
requests==2.25.1
starlette==0.19.0
starlette-exporter==0.12.0
uvicorn==0.17.6
jinja2==3.0.3
urllib3==1.26.9
pillow==9.0.1
httpx==0.23.0  # Required for TestClient
